debt(GQLV1): Remove legacy expense code
