+++ server/graphql/schemaV2.graphql
@@ -300,6 +300,14 @@ enum AccountType {
   PROJECT
 }
 
+input AccountUpdateInput {
+  """
+  The public id identifying the account (ie: dgm9bnk8-0437xqry-ejpvzeol-jdayw5re)
+  """
+  id: String!
+  currency: Currency
+}
+
 """
 An account that can receive financial contributions
 """
@@ -1080,7 +1088,6 @@ input CommentCreateInput {
   If your comment is linked to an expense, set it here
   """
   expense: ExpenseReferenceInput
-  ExpenseId: Int
   ConversationId: String
   update: UpdateReferenceInput
 }
@@ -3313,12 +3320,6 @@ input ExpenseCreateInput {
   payoutMethod: PayoutMethodInput!
 
   """
-  @deprecated 2020-04-08: Please use the items field - The list of items for
-  this expense. Total amount will be computed from them.
-  """
-  attachments: [ExpenseItemCreateInput]
-
-  """
   The list of items for this expense. Total amount will be computed from them.
   """
   items: [ExpenseItemCreateInput]
@@ -4426,6 +4427,11 @@ The name of the current plan and its characteristics.
 """
 type HostPlan {
   """
+  The public id identifying the account (ie: 5v08jk63-w4g9nbpz-j7qmyder-p7ozax5g)
+  """
+  id: String
+
+  """
   The name of the plan
   """
   name: String
@@ -5173,7 +5179,7 @@ type Mutation {
     A message to attach for the host to review the application
     """
     message: String
-  ): Collective!
+  ): Account!
 
   """
   Reply to a host application
@@ -5275,7 +5281,7 @@ type Mutation {
   """
   editHostPlan(
     """
-    Account that will have 2FA added to it
+    Account where the host plan will be edited.
     """
     account: AccountReferenceInput!
 
@@ -5283,6 +5289,16 @@ type Mutation {
     The name of the plan to subscribe to.
     """
     plan: String!
+  ): Host!
+
+  """
+  Edit key properties of an account.
+  """
+  editAccount(
+    """
+    Account to edit.
+    """
+    account: AccountUpdateInput!
   ): Host!
 
   """
@@ -5724,7 +5740,7 @@ type OrderWithPayment {
   order: Order!
 
   """
-  If donating as a guest, this will contain your guest token to contribute again in the future
+  If donating as a guest, this will contain your guest token to confirm your order
   """
   guestToken: String
