+++ server/graphql/v1/TransactionInterface.js
@@ -292,52 +292,12 @@ export const TransactionExpenseType = new GraphQLObjectType({
           return transaction.description || expense;
         },
       },
-      privateMessage: {
-        type: GraphQLString,
-        deprecationReason: 'Please use transaction.expense.privateMessage',
-        resolve(transaction, args, req) {
-          // If it's a expense transaction it'll have an ExpenseId
-          // otherwise we return null
-          return transaction.ExpenseId
-            ? req.loaders.Expense.byId.load(transaction.ExpenseId).then(expense => expense && expense.privateMessage)
-            : null;
-        },
-      },
-      category: {
-        type: GraphQLString,
-        deprecationReason: 'Please use transaction.expense.category',
-        resolve(transaction, args, req) {
-          // If it's a expense transaction it'll have an ExpenseId
-          // otherwise we return null
-          return transaction.ExpenseId
-            ? req.loaders.Expense.byId.load(transaction.ExpenseId).then(expense => expense && expense.tags?.[0])
-            : null;
-        },
-      },
       expense: {
         type: ExpenseType,
         resolve(transaction, args, req) {
           // If it's a expense transaction it'll have an ExpenseId
           // otherwise we return null
           return transaction.ExpenseId ? req.loaders.Expense.byId.load(transaction.ExpenseId) : null;
-        },
-      },
-      attachment: {
-        type: GraphQLString,
-        deprecationReason: 'Please use transaction.expense.attachment',
-        async resolve(transaction, args, req) {
-          // If it's a expense transaction it'll have an ExpenseId otherwise we return null
-          if (!transaction.ExpenseId) {
-            return null;
-          } else {
-            const expense = await req.loaders.Expense.byId.load(transaction.ExpenseId);
-            if (!expense || !(await canSeeExpenseAttachments(req, expense))) {
-              return null;
-            } else {
-              const items = await getExpenseItems(transaction.ExpenseId, req);
-              return items[0] && items[0].url;
-            }
-          }
         },
       },
     };
